'''Student's t-Distribution'''  * Code fragment from http://www.alglib.net/  * Testing work unit: TestStudentDistributionWorkUnit

+++ projects/Proteins/Rosetta/Data/Mover/PhiPsiMover.cpp
@@ -35,7 +35,7 @@ void PhiPsiMover::move(core::pose::PoseOP& pose, int residue, double deltaPhi, d
 void PhiPsiMover::move(PosePtr& pose, int residue, double deltaPhi, double deltaPsi)
 {
   jassert(isNumberValid(deltaPhi) && isNumberValid(deltaPsi));
-  jassert((residue >= 0) && (residue < pose->getLength()));
+  jassert((residue >= 0) && (residue < (int)pose->getLength()));
 
   pose->setPhi(residue, pose->getPhi(residue) + deltaPhi);
   pose->setPsi(residue, pose->getPsi(residue) + deltaPsi);
